refactor: create service class for request handling
