+++ src/utils/get.ts
@@ -5,10 +5,10 @@ import {
   invalidRequestUrlErrorHandler,
   invalidUserIdErrorHandler,
   notFoundErrorHandler,
-} from '../utils/errors.js';
-import { findUserIndex } from '../utils/user.js';
+} from './errors.js';
+import { findUserIndex } from './user.js';
 
-import { checkIfValidUUID } from '../utils/uuid.js';
+import { checkIfValidUUID } from './uuid.js';
 
 export function get(
   req: http.IncomingMessage,
